add: Production用のデータを作成する

+++ src/database/migrations/2024_10_05_030222_create_telescope_entries_table.php
@@ -1,8 +1,8 @@
 <?php
 
-use Illuminate\Database\Migrations\Migration;
-use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
 
 return new class extends Migration
 {
@@ -19,41 +19,43 @@ return new class extends Migration
      */
     public function up(): void
     {
-        $schema = Schema::connection($this->getConnection());
+        if (App::environment('local')) {
+            $schema = Schema::connection($this->getConnection());
 
-        $schema->create('telescope_entries', function (Blueprint $table) {
-            $table->bigIncrements('sequence');
-            $table->uuid('uuid');
-            $table->uuid('batch_id');
-            $table->string('family_hash')->nullable();
-            $table->boolean('should_display_on_index')->default(true);
-            $table->string('type', 20);
-            $table->longText('content');
-            $table->dateTime('created_at')->nullable();
+            $schema->create('telescope_entries', function (Blueprint $table) {
+                $table->bigIncrements('sequence');
+                $table->uuid('uuid');
+                $table->uuid('batch_id');
+                $table->string('family_hash')->nullable();
+                $table->boolean('should_display_on_index')->default(true);
+                $table->string('type', 20);
+                $table->longText('content');
+                $table->dateTime('created_at')->nullable();
 
-            $table->unique('uuid');
-            $table->index('batch_id');
-            $table->index('family_hash');
-            $table->index('created_at');
-            $table->index(['type', 'should_display_on_index']);
-        });
+                $table->unique('uuid');
+                $table->index('batch_id');
+                $table->index('family_hash');
+                $table->index('created_at');
+                $table->index(['type', 'should_display_on_index']);
+            });
 
-        $schema->create('telescope_entries_tags', function (Blueprint $table) {
-            $table->uuid('entry_uuid');
-            $table->string('tag');
+            $schema->create('telescope_entries_tags', function (Blueprint $table) {
+                $table->uuid('entry_uuid');
+                $table->string('tag');
 
-            $table->primary(['entry_uuid', 'tag']);
-            $table->index('tag');
+                $table->primary(['entry_uuid', 'tag']);
+                $table->index('tag');
 
-            $table->foreign('entry_uuid')
-                ->references('uuid')
-                ->on('telescope_entries')
-                ->onDelete('cascade');
-        });
+                $table->foreign('entry_uuid')
+                    ->references('uuid')
+                    ->on('telescope_entries')
+                    ->onDelete('cascade');
+            });
 
-        $schema->create('telescope_monitoring', function (Blueprint $table) {
-            $table->string('tag')->primary();
-        });
+            $schema->create('telescope_monitoring', function (Blueprint $table) {
+                $table->string('tag')->primary();
+            });
+        }
     }
 
     /**
